+++ include/tern/user.h
@@ -20,12 +20,12 @@ extern "C" {
 
   /// New programming primitives to get better performance without affecting
   /// logics of applications.
-  void tern_lineup_init(unsigned opaque_type, unsigned count, unsigned timeout_turns);
-  void tern_lineup_start(unsigned opaque_type);
-  void tern_lineup_end(unsigned opaque_type);
+  void tern_lineup_init(long opaque_type, unsigned count, unsigned timeout_turns);
+  void tern_lineup_start(long opaque_type);
+  void tern_lineup_end(long opaque_type);
 
-  void tern_workload_start(unsigned opaque_type, unsigned workload_hint);
-  void tern_workload_end(unsigned opaque_type);
+  void tern_workload_start(long opaque_type, unsigned workload_hint);
+  void tern_workload_end(long opaque_type);
 
   void tern_non_det_start();
   void tern_non_det_end();
